remove set shake slideout logic
