+++ app/src/organisms/ModuleCard/HeaterShakerSlideout.tsx
@@ -1,16 +1,12 @@
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
-import { useSelector } from 'react-redux'
 import {
   useCreateCommandMutation,
   useCreateLiveCommandMutation,
 } from '@opentrons/react-api-client'
 import {
   getModuleDisplayName,
-  RPM,
   CELSIUS,
-  HS_RPM_MAX,
-  HS_RPM_MIN,
   HS_TEMP_MIN,
   HS_TEMP_MAX,
 } from '@opentrons/shared-data'
@@ -23,27 +19,19 @@ import {
   SPACING,
   Text,
   TYPOGRAPHY,
-  useConditionalConfirm,
 } from '@opentrons/components'
-import { getIsHeaterShakerAttached } from '../../redux/config'
 import { InputField } from '../../atoms/InputField'
-import { Portal } from '../../App/portal'
 import { SubmitPrimaryButton } from '../../atoms/buttons'
 import { useRunStatuses } from '../Devices/hooks'
-import { ConfirmAttachmentModal } from './ConfirmAttachmentModal'
 import { useModuleIdFromRun } from './useModuleIdFromRun'
 
 import type { HeaterShakerModule } from '../../redux/modules/types'
-import type {
-  HeaterShakerSetAndWaitForShakeSpeedCreateCommand,
-  HeaterShakerStartSetTargetTemperatureCreateCommand,
-} from '@opentrons/shared-data/protocol/types/schemaV6/command/module'
+import type { HeaterShakerStartSetTargetTemperatureCreateCommand } from '@opentrons/shared-data/protocol/types/schemaV6/command/module'
 
 interface HeaterShakerSlideoutProps {
   module: HeaterShakerModule
   onCloseClick: () => unknown
   isExpanded: boolean
-  isSetShake: boolean
   isLoadedInRun: boolean
   currentRunId?: string
 }
@@ -55,7 +43,6 @@ export const HeaterShakerSlideout = (
     module,
     onCloseClick,
     isExpanded,
-    isSetShake,
     isLoadedInRun,
     currentRunId,
   } = props
@@ -65,12 +52,11 @@ export const HeaterShakerSlideout = (
   const { isRunIdle, isRunTerminal } = useRunStatuses()
   const { createCommand } = useCreateCommandMutation()
   const moduleName = getModuleDisplayName(module.moduleModel)
-  const configHasHeaterShakerAttached = useSelector(getIsHeaterShakerAttached)
   const { moduleIdFromRun } = useModuleIdFromRun(
     module,
     currentRunId != null ? currentRunId : null
   )
-  const modulePart = isSetShake ? t('shake_speed') : t('temperature')
+  const modulePart = t('temperature')
 
   let moduleId: string
   if (isRunIdle && currentRunId != null && isLoadedInRun) {
@@ -79,51 +65,11 @@ export const HeaterShakerSlideout = (
     moduleId = module.id
   }
 
-  const sendShakeSpeedCommand = (): void => {
-    if (hsValue != null && isSetShake) {
-      const setShakeCommand: HeaterShakerSetAndWaitForShakeSpeedCreateCommand = {
-        commandType: 'heaterShaker/setAndWaitForShakeSpeed',
-        params: {
-          moduleId: moduleId,
-          rpm: hsValue,
-        },
-      }
-      if (isRunIdle && currentRunId != null && isLoadedInRun) {
-        createCommand({ runId: currentRunId, command: setShakeCommand }).catch(
-          (e: Error) => {
-            console.error(
-              `error setting heater shaker shake speed: ${e.message} with run id ${currentRunId}`
-            )
-          }
-        )
-      } else if (
-        (currentRunId != null && isRunTerminal) ||
-        currentRunId == null
-      ) {
-        createLiveCommand({
-          command: setShakeCommand,
-        }).catch((e: Error) => {
-          console.error(`error setting heater shaker shake speed: ${e.message}`)
-        })
-      }
-    }
-    onCloseClick()
-    setHsValue(null)
-  }
-  const {
-    confirm: confirmAttachment,
-    showConfirmation: showConfirmationModal,
-    cancel: cancelExit,
-  } = useConditionalConfirm(
-    sendShakeSpeedCommand,
-    !configHasHeaterShakerAttached
-  )
-
   const sendSetTemperatureOrShakeCommand: React.MouseEventHandler<HTMLInputElement> = e => {
     e.preventDefault()
     e.stopPropagation()
 
-    if (hsValue != null && !isSetShake) {
+    if (hsValue != null) {
       const setTempCommand: HeaterShakerStartSetTargetTemperatureCreateCommand = {
         commandType: 'heaterShaker/setTargetTemperature',
         params: {
@@ -152,30 +98,18 @@ export const HeaterShakerSlideout = (
         })
       }
     }
-    if (isSetShake) {
-      confirmAttachment()
-    } else {
-      setHsValue(null)
-      onCloseClick()
-    }
+    setHsValue(null)
+    onCloseClick()
   }
 
-  let errorMessage
-  if (isSetShake) {
-    errorMessage =
-      hsValue != null && (hsValue < HS_RPM_MIN || hsValue > HS_RPM_MAX)
-        ? t('input_out_of_range')
-        : null
-  } else {
-    errorMessage =
-      hsValue != null && (hsValue < HS_TEMP_MIN || hsValue > HS_TEMP_MAX)
-        ? t('input_out_of_range')
-        : null
-  }
+  const errorMessage =
+    hsValue != null && (hsValue < HS_TEMP_MIN || hsValue > HS_TEMP_MAX)
+      ? t('input_out_of_range')
+      : null
 
-  const inputMax = isSetShake ? HS_RPM_MAX : HS_TEMP_MAX
-  const inputMin = isSetShake ? HS_RPM_MIN : HS_TEMP_MIN
-  const unit = isSetShake ? RPM : CELSIUS
+  const inputMax = HS_TEMP_MAX
+  const inputMin = HS_TEMP_MIN
+  const unit = CELSIUS
 
   const handleCloseSlideout = (): void => {
     setHsValue(null)
@@ -184,15 +118,6 @@ export const HeaterShakerSlideout = (
 
   return (
     <>
-      {showConfirmationModal && (
-        <Portal level="top">
-          <ConfirmAttachmentModal
-            onCloseClick={cancelExit}
-            isProceedToRunModal={false}
-            onConfirmClick={sendShakeSpeedCommand}
-          />
-        </Portal>
-      )}
       <Slideout
         title={t('set_status_heater_shaker', {
           part: modulePart,
@@ -216,7 +141,7 @@ export const HeaterShakerSlideout = (
           paddingTop={SPACING.spacing2}
           data-testid={`HeaterShakerSlideout_title_${module.serialNumber}`}
         >
-          {isSetShake ? t('set_shake_of_hs') : t('set_target_temp_of_hs')}
+          {t('set_target_temp_of_hs')}
         </Text>
         <Flex
           marginTop={SPACING.spacing4}
@@ -229,12 +154,12 @@ export const HeaterShakerSlideout = (
             color={COLORS.darkGrey}
             marginBottom={SPACING.spacing3}
           >
-            {isSetShake ? t('set_shake_speed') : t('set_block_temp')}
+            {t('set_block_temp')}
           </Text>
           <form id="HeaterShakerSlideout_submitValue">
             <InputField
-              data-testid={`${module.moduleModel}_${isSetShake}`}
-              id={`${module.moduleModel}_${isSetShake}`}
+              data-testid={`${module.moduleModel}_setTemp`}
+              id={`${module.moduleModel}_setTemp`}
               units={unit}
               autoFocus
               value={hsValue != null ? Math.round(hsValue) : null}
