+++ app/src/organisms/ModuleCard/__tests__/HeaterShakerSlideout.test.tsx
@@ -6,30 +6,20 @@ import {
 } from '@opentrons/react-api-client'
 import { fireEvent } from '@testing-library/react'
 import { i18n } from '../../../i18n'
-import { getIsHeaterShakerAttached } from '../../../redux/config'
 import { mockHeaterShaker } from '../../../redux/modules/__fixtures__'
 import { useRunStatuses } from '../../Devices/hooks'
 import { useModuleIdFromRun } from '../useModuleIdFromRun'
 import { HeaterShakerSlideout } from '../HeaterShakerSlideout'
-import { ConfirmAttachmentModal } from '../ConfirmAttachmentModal'
 
 jest.mock('@opentrons/react-api-client')
-jest.mock('../ConfirmAttachmentModal')
-jest.mock('../../../redux/config')
 jest.mock('../useModuleIdFromRun')
 jest.mock('../../Devices/hooks')
 
-const mockGetIsHeaterShakerAttached = getIsHeaterShakerAttached as jest.MockedFunction<
-  typeof getIsHeaterShakerAttached
->
 const mockUseLiveCommandMutation = useCreateLiveCommandMutation as jest.MockedFunction<
   typeof useCreateLiveCommandMutation
 >
 const mockUseCommandMutation = useCreateCommandMutation as jest.MockedFunction<
   typeof useCreateCommandMutation
->
-const mockConfirmAttachmentModal = ConfirmAttachmentModal as jest.MockedFunction<
-  typeof ConfirmAttachmentModal
 >
 const mockUseModuleIdFromRun = useModuleIdFromRun as jest.MockedFunction<
   typeof useModuleIdFromRun
@@ -58,7 +48,6 @@ describe('HeaterShakerSlideout', () => {
       isRunTerminal: true,
       isRunIdle: false,
     })
-    mockGetIsHeaterShakerAttached.mockReturnValue(false)
     mockUseLiveCommandMutation.mockReturnValue({
       createLiveCommand: mockCreateLiveCommand,
     } as any)
@@ -69,9 +58,6 @@ describe('HeaterShakerSlideout', () => {
       createCommand: mockCreateCommand,
     } as any)
 
-    mockConfirmAttachmentModal.mockReturnValue(
-      <div>mock confirm attachment modal</div>
-    )
     mockUseModuleIdFromRun.mockReturnValue({
       moduleIdFromRun: 'heatershaker_id',
     })
@@ -81,25 +67,9 @@ describe('HeaterShakerSlideout', () => {
     jest.resetAllMocks()
   })
 
-  it('renders correct title and body for Heater shaker set shake', () => {
-    props = {
-      module: mockHeaterShaker,
-      isSetShake: true,
-      isExpanded: true,
-      onCloseClick: jest.fn(),
-      isLoadedInRun: false,
-    }
-    const { getByText } = render(props)
-
-    getByText('Set shake speed for Heater-Shaker Module GEN1')
-    getByText('Set rpm for this module.')
-    getByText('Confirm')
-  })
-
   it('renders correct title and body for heatershaker set temperature', () => {
     props = {
       module: mockHeaterShaker,
-      isSetShake: false,
       isExpanded: true,
       onCloseClick: jest.fn(),
       isLoadedInRun: false,
@@ -113,35 +83,16 @@ describe('HeaterShakerSlideout', () => {
     getByText('Confirm')
   })
 
-  it('renders the button and it is not clickable until there is something in form field for set shake', () => {
-    props = {
-      module: mockHeaterShaker,
-      isSetShake: true,
-      isExpanded: true,
-      onCloseClick: jest.fn(),
-      isLoadedInRun: false,
-    }
-
-    const { getByRole, getByTestId, getByText } = render(props)
-    const button = getByRole('button', { name: 'Confirm' })
-    const input = getByTestId('heaterShakerModuleV1_true')
-    fireEvent.change(input, { target: { value: '300' } })
-    expect(button).toBeEnabled()
-    fireEvent.click(button)
-    getByText('mock confirm attachment modal')
-  })
-
   it('renders the button and it is not clickable until there is something in form field for set temp', () => {
     props = {
       module: mockHeaterShaker,
-      isSetShake: false,
       isExpanded: true,
       onCloseClick: jest.fn(),
       isLoadedInRun: false,
     }
     const { getByRole, getByTestId } = render(props)
     const button = getByRole('button', { name: 'Confirm' })
-    const input = getByTestId('heaterShakerModuleV1_false')
+    const input = getByTestId('heaterShakerModuleV1_setTemp')
     fireEvent.change(input, { target: { value: '40' } })
     expect(button).toBeEnabled()
     fireEvent.click(button)
@@ -161,63 +112,18 @@ describe('HeaterShakerSlideout', () => {
   it('renders the exit button and when clicked, deletes the value input', () => {
     props = {
       module: mockHeaterShaker,
-      isSetShake: false,
       isExpanded: true,
       onCloseClick: jest.fn(),
       isLoadedInRun: false,
     }
     const { getByLabelText, getByTestId } = render(props)
     const button = getByLabelText('exit')
-    const input = getByTestId('heaterShakerModuleV1_false')
+    const input = getByTestId('heaterShakerModuleV1_setTemp')
     fireEvent.change(input, { target: { value: '40' } })
     fireEvent.click(button)
 
     expect(props.onCloseClick).toHaveBeenCalled()
     expect(input).not.toHaveValue()
-  })
-
-  it('renders heater shaker form field and when button is clicked, confirm attachment modal is not rendered', () => {
-    mockGetIsHeaterShakerAttached.mockReturnValue(true)
-    props = {
-      module: mockHeaterShaker,
-      isSetShake: true,
-      isExpanded: true,
-      onCloseClick: jest.fn(),
-      isLoadedInRun: false,
-    }
-
-    const { getByRole, getByTestId } = render(props)
-    const button = getByRole('button', { name: 'Confirm' })
-    const input = getByTestId('heaterShakerModuleV1_true')
-    fireEvent.change(input, { target: { value: '300' } })
-    expect(button).toBeEnabled()
-    fireEvent.click(button)
-    expect(props.onCloseClick).toHaveBeenCalled()
-  })
-
-  it('renders the button and it is not clickable until there is something in form field for set shake when there is a runId', () => {
-    mockUseRunStatuses.mockReturnValue({
-      isLegacySessionInProgress: false,
-      isRunStill: false,
-      isRunTerminal: false,
-      isRunIdle: true,
-    })
-    props = {
-      module: mockHeaterShaker,
-      isSetShake: true,
-      isExpanded: true,
-      onCloseClick: jest.fn(),
-      isLoadedInRun: true,
-      currentRunId: 'test123',
-    }
-
-    const { getByRole, getByTestId, getByText } = render(props)
-    const button = getByRole('button', { name: 'Confirm' })
-    const input = getByTestId('heaterShakerModuleV1_true')
-    fireEvent.change(input, { target: { value: '300' } })
-    expect(button).toBeEnabled()
-    fireEvent.click(button)
-    getByText('mock confirm attachment modal')
   })
 
   it('renders the button and it is not clickable until there is something in form field for set temp when there is a runId', () => {
@@ -229,7 +135,6 @@ describe('HeaterShakerSlideout', () => {
     })
     props = {
       module: mockHeaterShaker,
-      isSetShake: false,
       isExpanded: true,
       onCloseClick: jest.fn(),
       isLoadedInRun: true,
@@ -237,7 +142,7 @@ describe('HeaterShakerSlideout', () => {
     }
     const { getByRole, getByTestId } = render(props)
     const button = getByRole('button', { name: 'Confirm' })
-    const input = getByTestId('heaterShakerModuleV1_false')
+    const input = getByTestId('heaterShakerModuleV1_setTemp')
     fireEvent.change(input, { target: { value: '40' } })
     expect(button).toBeEnabled()
     fireEvent.click(button)
